Extract SQLite source safety helpers
